+++ CHANGELOG.md
@@ -5,3 +5,6 @@
 - Added Docker primer (docker/notes/0000-primer-docker.md)
 - Added Docker install script (docker/scripts/2026-06-16-install-docker.sh)
 - Added Docker CLI exploration notes (docker/notes/2026-06-16-explore-docker-cli.md)
+- Added Git primer (git/notes/0000-primer-git.md)
+- Added Git configure & first commit script (git/scripts/configure-and-commit-first.sh)
+- Added Git CLI exploration notes (git/notes/2026-06-16-explore-git-cli.md)